backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
jinja2>=3.1.2
werkzeug>=3.0.1
